+++ map-structure/vi-map/include/vi-map/vertex.h
@@ -88,7 +88,7 @@ class Vertex : public pose_graph::Vertex {
       const aslam::VisualFrame::DescriptorsT& descriptors,
       const Eigen::VectorXd& descriptor_scales,
       const std::vector<LandmarkId>& observed_landmark_ids,
-      const Eigen::Matrix4Xi& semantic_object_measurements,
+      const Eigen::Matrix4Xd& semantic_object_measurements,
       const Eigen::VectorXd& semantic_object_uncertainties,
       const Eigen::VectorXi& semantic_object_class_ids,
       const aslam::VisualFrame::SemanticObjectDescriptorsT&
@@ -104,7 +104,7 @@ class Vertex : public pose_graph::Vertex {
       const Eigen::VectorXd& uncertainties,
       const aslam::VisualFrame::DescriptorsT& descriptors,
       const std::vector<LandmarkId>& observed_landmark_ids,
-      const Eigen::Matrix4Xi& semantic_object_measurements,
+      const Eigen::Matrix4Xd& semantic_object_measurements,
       const Eigen::VectorXd& semantic_object_uncertainties,
       const Eigen::VectorXi& semantic_object_class_ids,
       const aslam::VisualFrame::SemanticObjectDescriptorsT&
